♻️  zb: Split handshake's command code into its own mod
